requirements.txt
pymongo==3.5.1
python-dateutil==2.6.1
Scrapy==1.4.0
service-identity==17.0.0
